improve performance and move item-related checks to utils
