+++ trafilatura/utils.py
@@ -468,7 +468,7 @@ def is_in_table_cell(elem: _Element) -> bool:
     # return elem.getparent() is not None and bool(elem.xpath('//ancestor::cell'))
     if elem.getparent() is None:
         return False
-    current = elem
+    current: Optional[_Element] = elem
     while current is not None:
         if current.tag == 'cell':
             return True
@@ -488,3 +488,49 @@ def is_last_element_in_cell(elem: _Element) -> bool:
         parent = cast(_Element, elem.getparent())
         children = parent.getchildren()
         return not children or children[-1] == elem
+
+
+def is_element_in_item(element: _Element) -> bool:
+    """Check whether an element is a list item or within a list item"""
+    current: Optional[_Element] = element
+    while current is not None:
+        if current.tag == 'item':
+            return True
+        current = current.getparent()
+    return False
+
+
+def is_first_element_in_item(element: _Element) -> bool:
+    """Check whether an element is the first element in list item"""
+    if element.tag == 'item' and element.text:
+        return True
+
+    current: Optional[_Element] = element
+    item_ancestor = None
+    while current is not None:
+        if current.tag == 'item':
+            item_ancestor = current
+            break
+        current = current.getparent()
+
+    if item_ancestor is None:
+        return False
+    elif not item_ancestor.text:
+        return True
+    return False
+
+
+def is_last_element_in_item(element: _Element) -> bool:
+    """Check whether an element is the last element in list item"""
+    if not is_element_in_item(element):
+        return False
+
+    # pure text only in list item
+    if element.tag == 'item':
+        return len(element.getchildren()) == 0
+    # element within list item
+    next_element = element.getnext()
+    if next_element is None:
+        return True
+    else:
+        return next_element.tag == 'item'
